create separate component for calendar animation
